Update timer method, change menu to update via msg
The timer method change was needed 1) because this is simpler to follow
and 2) because it is easier to control incrementing the minutes field of
the CountdownTimer.

The menu needed to be updated only via sending messages to the
background for a smoother UI experience. Otherwise it would be set to a
default value and then quickly switch to whatever the timer was at.